# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cafe Supply Chain API",
        "description": "REST API documentation split by platform: Admin Dashboard, Cafe Mobile App, and Public Storefront.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/purchase-orders": {
            "get": {
                "tags": [
                    "Admin Inventory"
                ],
                "summary": "List purchase orders",
                "operationId": "3fdefa42a6d5f9f20c6ffe0ea3f2b802",
                "responses": {
                    "200": {
                        "description": "Paginated list of purchase orders"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin Inventory"
                ],
                "summary": "Create a purchase order",
                "operationId": "a3fcacac403b00a3481d05a6ede6f362",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase order created"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/purchase-orders/{id}": {
            "get": {
                "tags": [
                    "Admin Inventory"
                ],
                "summary": "Get a purchase order",
                "operationId": "2862f9933067e46980cb89f44c6d2d4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase order details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin Inventory"
                ],
                "summary": "Update a purchase order",
                "operationId": "705daf8039aec99fe1b173889855aeb5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Purchase order updated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin Inventory"
                ],
                "summary": "Delete a purchase order",
                "operationId": "eaf7cd7ce602270f2039c1b44c213550",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Purchase order deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PurchaseOrderRequest": {
                "required": [
                    "supplier_id",
                    "warehouse_id",
                    "status"
                ],
                "properties": {
                    "supplier_id": {
                        "type": "integer"
                    },
                    "warehouse_id": {
                        "type": "integer"
                    },
                    "order_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "string",
                        "maxLength": 20
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}